move the get tweet into twitter
